Refactored Cryptomator Intents
- added `VaultOptionsProvider` to handle provideVaultOptions calls across the IntentHandlers
- moved `oneTimeSetup()` to the main `IntentHandler`
