+++ CryptomatorIntents/LockVaultIntentHandler.swift
@@ -13,17 +13,10 @@ import Intents
 import Promises
 
 class LockVaultIntentHandler: NSObject, LockVaultIntentHandling {
-	private let vaultAccountManager: VaultAccountManager
-	private let cloudProviderAccountManager: CloudProviderAccountManager
+	let vaultOptionsProvider: VaultOptionsProvider
 
-	override convenience init() {
-		LockVaultIntentHandler.oneTimeSetup()
-		self.init(vaultAccountManager: VaultAccountDBManager.shared, cloudProviderAccountManager: CloudProviderAccountDBManager.shared)
-	}
-
-	init(vaultAccountManager: VaultAccountManager, cloudProviderAccountManager: CloudProviderAccountManager) {
-		self.vaultAccountManager = vaultAccountManager
-		self.cloudProviderAccountManager = cloudProviderAccountManager
+	init(vaultOptionsProvider: VaultOptionsProvider) {
+		self.vaultOptionsProvider = vaultOptionsProvider
 	}
 
 	func handle(intent: LockVaultIntent) async -> LockVaultIntentResponse {
@@ -41,44 +34,15 @@ class LockVaultIntentHandler: NSObject, LockVaultIntentHandling {
 
 	@available(iOS, introduced: 13.0, deprecated: 14.0, message: "")
 	func provideVaultOptions(for intent: LockVaultIntent, with completion: @escaping ([Vault]?, Error?) -> Void) {
-		do {
-			let vaultAccounts = try vaultAccountManager.getAllAccounts()
-			let vaults: [Vault] = vaultAccounts.map {
-				return Vault(identifier: $0.vaultUID, display: $0.vaultName)
-			}
-			completion(vaults, nil)
-		} catch {
-			completion(nil, error)
-		}
+		vaultOptionsProvider.provideVaultOptions(with: completion)
 	}
 
 	@available(iOSApplicationExtension 14.0, *)
 	func provideVaultOptionsCollection(for intent: LockVaultIntent) async throws -> INObjectCollection<Vault> {
-		let vaultAccounts = try vaultAccountManager.getAllAccounts()
-		let vaults: [Vault] = try vaultAccounts.map {
-			let cloudProviderType = try cloudProviderAccountManager.getCloudProviderType(for: $0.delegateAccountUID)
-			return Vault(identifier: $0.vaultUID,
-			             display: $0.vaultName,
-			             subtitle: $0.vaultPath.path,
-			             image: .init(type: cloudProviderType))
-		}
-		return INObjectCollection(items: vaults)
+		return try await vaultOptionsProvider.provideVaultOptionsCollection()
 	}
 
 	// MARK: Internal
-
-	private static var oneTimeSetup: () -> Void = {
-		// Set up logger
-		LoggerSetup.oneTimeSetup()
-		if let dbURL = CryptomatorDatabase.sharedDBURL {
-			do {
-				let dbPool = try CryptomatorDatabase.openSharedDatabase(at: dbURL)
-				CryptomatorDatabase.shared = try CryptomatorDatabase(dbPool)
-			} catch {
-				DDLogError("Open shared database at \(dbURL) failed with error: \(error)")
-			}
-		}
-	}
 
 	private func lockVault(with domainIdentifier: NSFileProviderDomainIdentifier) async throws {
 		let getXPCPromise: Promise<XPC<VaultLocking>> = FileProviderXPCConnector.shared.getXPC(serviceName: .vaultLocking, domainIdentifier: domainIdentifier)
